+++ R/postestimate_assess.R
@@ -80,14 +80,14 @@
 #' \item{Degrees of freedom, "df"}{
 #'   Returns the degrees of freedom. Calculation is done by [calculateDf()].
 #'   }
-#' \item{Effect size; "esize"}{An index of the effect size of an independent
-#'   variable in a structural regression equation. The effect size of the k'th
+#' \item{Effect size; "f2"}{An index of the effect size of an independent
+#'   variable in a structural regression equation. This measure is commonly 
+#'   known as Cohen's f^2. The effect size of the k'th
 #'   independent variable in this case
 #'   is definied as the ratio (R2_included - R2_excluded)/(1 - R2_included), where 
 #'   R2_included and R2_excluded are the R squares of the 
 #'   original structural model regression equation (R2_included) and the
 #'   alternative specification with the k'th variable dropped (R2_excluded).
-#'   This measure is commonly known as Cohen's f^2.
 #'   Calculation is done by [calculateEffectSize()].}
 #' \item{Fit indices; "chi_square", "chi_square_df", "cfi", "gfi", "ifi", "nfi", 
 #'       "nnfi",  "rmsea", "rms_theta", "rms_theta_mi", "srmr"}{
@@ -202,7 +202,7 @@
 #'   .quality_criterion   = c("all", "ave", "rho_C", "rho_C_mm", "rho_C_weighted", 
 #'                            "rho_C_weighted_mm", "cronbachs_alpha", 
 #'                           "cronbachs_alpha_weighted", "dg", "dl", "dml", "df",
-#'                           "esize", "chi_square", "chi_square_df",
+#'                           "f2", "chi_square", "chi_square_df",
 #'                           "cfi", "gfi", "ifi", "nfi", "nnfi", 
 #'                           "reliability", 
 #'                           "rmsea", "rms_theta", "rms_theta_mi", "srmr",
@@ -231,7 +231,7 @@ assess <- function(
   .quality_criterion   = c("all", "ave", "rho_C", "rho_C_mm", "rho_C_weighted", 
                            "rho_C_weighted_mm", "cronbachs_alpha", 
                            "cronbachs_alpha_weighted", "dg", "dl", "dml", "df",
-                           "esize", "chi_square", "chi_square_df",
+                           "f2", "chi_square", "chi_square_df",
                            "cfi", "gfi", "ifi", "nfi", "nnfi", 
                            "reliability",
                            "rmsea", "rms_theta", "rms_theta_mi", "srmr",
@@ -251,7 +251,7 @@ assess.cSEMResults_default <- function(
   .quality_criterion   = c("all", "ave", "rho_C", "rho_C_mm", "rho_C_weighted", 
                            "rho_C_weighted_mm", "cronbachs_alpha", 
                            "cronbachs_alpha_weighted", "dg", "dl", "dml", "df",
-                           "esize", "chi_square", "chi_square_df",
+                           "f2", "chi_square", "chi_square_df",
                            "cfi", "gfi", "ifi", "nfi", "nnfi", 
                            "reliability",
                            "rmsea", "rms_theta", "rms_theta_mi", "srmr",
@@ -329,9 +329,9 @@ assess.cSEMResults_default <- function(
     # dML
     out[["Df"]]   <- calculateDf(.object, ...)
   }
-  if(any(.quality_criterion %in% c("all", "esize")) && !all(m$structural == 0)) {
-    # Effect size
-    out[["Effect_size"]] <- calculateEffectSize(.object)
+  if(any(.quality_criterion %in% c("all", "f2")) && !all(m$structural == 0)) {
+    # Effect size (f2)
+    out[["F2"]] <- calculatef2(.object)
   }
   if(any(.quality_criterion %in% c("all", "chi_square"))) {
     # Effect size
@@ -508,7 +508,7 @@ assess.cSEMResults_multi <- function(
   .quality_criterion   = c("all", "ave", "rho_C", "rho_C_mm", "rho_C_weighted", 
                            "rho_C_weighted_mm", "cronbachs_alpha",  
                            "cronbachs_alpha_weighted", "dg", "dl", "dml", "df",
-                           "esize", "chi_square", "chi_square_df",
+                           "f2", "chi_square", "chi_square_df",
                            "cfi", "gfi", "ifi", "nfi", "nnfi", 
                            "reliability",
                            "rmsea", "rms_theta", "rms_theta_mi", "srmr",
@@ -541,7 +541,7 @@ assess.cSEMResults_2ndorder <- function(
   .quality_criterion   = c("all", "ave", "rho_C", "rho_C_mm", "rho_C_weighted", 
                            "rho_C_weighted_mm", "cronbachs_alpha",  
                            "cronbachs_alpha_weighted", "dg", "dl", "dml", "df",
-                           "esize", "chi_square", "chi_square_df",
+                           "f2", "chi_square", "chi_square_df",
                            "cfi", "gfi", "ifi", "nfi", "nnfi", 
                            "reliability",
                            "rmsea", "rms_theta", "rms_theta_mi", "srmr",